Random and Update AI
Added a Random AI that changes direction each time interval. Time
interval is determined by a random number between 5 and 20.

Updated follow logic, but it is not currently clean because of
collisions. The collisions cause a delay on the player character
movement, which causes the enemy to effectively trap the player at walls
by continuously moving towards them.